% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{AggregatedCoverage-class}
\alias{AggregatedCoverage-class}
\title{AggregatedCoverage class}
\description{
This class is based on the \code{SummarizedExperiment} class. It implements
bigwig file import over a set of \code{GRanges} of interest and computes
three metrics: 1) mean coverage, 2) lower confidence interval (CI) and 3)
upper CI boundaries. It stores these metrics in a rectangular assay, where
rows represent distance to the center of \code{GRanges} of interest, and columns
correspond to individual bigwig coverage tracks
}
