% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{AggregatedCoverage}
\alias{AggregatedCoverage}
\title{AggregatedCoverage constructor}
\usage{
AggregatedCoverage(granges, colData)
}
\arguments{
\item{granges}{Genomic loci of interest provided as a \code{GRanges} object.
All the loci should have the same width.}

\item{colData}{Data frame with a \code{file} column, containing the path
to individual \code{bigwig} files. Additional metadata columns can be provided.}
}
\value{
an \code{AggregatedCoverage} object
}
\description{
AggregatedCoverage constructor
}
\examples{
files <- data.frame(file = c(
  system.file('extdata', 'Scc1-calibrated_rep1.bw', package = 'aggregatR'),
  system.file('extdata', 'Scc1-calibrated_rep2.bw', package = 'aggregatR')
))
data(Scc1_peaks)
AC <- AggregatedCoverage(Scc1_peaks, files)
AC
}
