% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{.check_widths}
\alias{.check_widths}
\title{.check_widths}
\usage{
.check_widths(granges)
}
\arguments{
\item{granges}{\code{GRanges} object}
}
\value{
width of the elements within the provided \code{GRanges} object
}
\description{
Internal check function
}
