% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{.filterGRanges}
\alias{.filterGRanges}
\title{.filterGRanges}
\usage{
.filterGRanges(l)
}
\arguments{
\item{l}{named list of 2 (coverage and features)}
}
\value{
named list of 2 (coverage and features)
}
\description{
This function takes a named list of 2 (coverage and features) and filters
out the features which are not strictly contained within the coverage track.
}
