% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{.parseBigWig}
\alias{.parseBigWig}
\title{.parseBigWig}
\usage{
.parseBigWig(file, granges)
}
\arguments{
\item{file}{Path to a \code{bigwig} file}

\item{granges}{\code{GRanges} object}
}
\value{
Data frame with columns containing average coverage and its lower and upper CI
boundaries around the center of \code{GRanges} provided as input.
}
\description{
Internal parsing function
}
