% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.prepareRowData}
\alias{.prepareRowData}
\title{.prepareRowData}
\usage{
.prepareRowData(width)
}
\arguments{
\item{width}{a single numeric value corresponding to the width of \code{GRanges}
used in the parent function.}
}
\value{
\code{data.frame} with a single \code{pos} column indicating the distance to
the center of \code{GRanges} width.
}
\description{
Internal utils function
}
