% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot,AggregatedCoverage,ANY-method}
\alias{plot,AggregatedCoverage,ANY-method}
\title{AggregateCoverage method for plot generic}
\usage{
\S4method{plot}{AggregatedCoverage,ANY}(x)
}
\arguments{
\item{x}{an AggregatedCoverage object}
}
\description{
This plot method is designed for \code{AggregatedCoverage} objects. It returns a
\code{ggplot} plot which can be further customized using \code{ggplot2} functions.
}
\examples{
files <- data.frame(file = c(
  system.file('extdata', 'Scc1-calibrated_rep1.bw', package = 'aggregatR'),
  system.file('extdata', 'Scc1-calibrated_rep2.bw', package = 'aggregatR')
))
data(Scc1_peaks)
AC <- AggregatedCoverage(Scc1_peaks, files)
plot(AC)
}
