% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Scc1_peaks}
\alias{Scc1_peaks}
\title{Scc1_peaks}
\format{
An object of class \code{GRanges} of length 1340.
}
\usage{
Scc1_peaks
}
\description{
Scc1 ChIP-seq peaks annotated by MACS2.
All the peaks have been extended to 5kb around their center.
}
\keyword{datasets}
